judge_legacy/requirements.txt
fastapi>=0.104.0
uvicorn>=0.30.0
pydantic>=2.5.0
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
ollama>=0.2.0
python-dotenv>=1.0.0